app/cars/serializers.py
from rest_framework.serializers import ModelSerializer

from cars.models import Car


class CarSerializer(ModelSerializer):
    class Meta:
        model = Car
        fields = ('current_location', 'number', 'lifting_capacity')
