requirements.txt
Django==4.2.11
djangorestframework==3.15.0
psycopg2==2.8.6
geopy==2.4.1